requirements.txt

requests
twilio
PyYAML